Fixed Dashboard-route, project restructuring, +++ * Made the Dashboard-route a protected route, like intended. This presents a redirecting loop when logged in. * Removed a redundant "Dashboard"-page since the "DashboardPage"-file was moved to the "pages"-folder. * Changed file-import paths to reflect new locations for pages and components. * Restructured the project, moved pages to the "pages"-folder. Created "styles"-folder for components and pages, moved css-files there. * Changed imports of CSS-files to direct to new routes. * Updated package-lock.json-files in frontend with a new frontend version, new package versions, and audited/fixed dependencies.
