+++ frontend/src/components/styles/components.css
