+++ frontend/src/pages/DashboardPage.jsx
@@ -1,11 +1,63 @@
-import React from 'react';
+import React, { useEffect, useState } from "react";
+import { Link } from "react-router-dom";
+import CreateStudyButton from "../components/Buttons/CreateStudyButton";
+import dummyStudies from "./DummyStudies";
+import "./styles/DashboardPage.css";
 
 const Dashboard = () => {
-    return (
-        <div>
-            <p>dashboard</p>
-        </div>
-    );
+  const [studies, setStudies] = useState([]);
+
+  // Load dummy studies on mount
+  useEffect(() => {
+    setStudies(dummyStudies);
+  }, []);
+
+  return (
+    <div className="DashboardWrapper">
+      <div className="DashboardHeader">
+        <h1>Dashboard</h1>
+        <Link to="/create-study" className="DashboardButton">
+          Create Study
+        </Link>
+      </div>
+      <div className="DashboardTable">
+        <table>
+          <thead>
+            <tr>
+              <th>Study</th>
+              <th>Created</th>
+              <th>Respondents</th>
+              <th>Results</th>
+              <th>Edit</th>
+              <th>Status</th>
+            </tr>
+          </thead>
+          <tbody>
+            {studies.map((study, index) => (
+              <tr key={index}>
+                <td>{study.study}</td>
+                <td>{study.created}</td>
+                <td>{study.respondents}</td>
+                <td>
+                  <a href={study.results} className="view-link">
+                    View
+                  </a>
+                </td>
+                <td>
+                  {study.status === "Active" && (
+                    <a href={study.edit} className="edit-link">
+                      Edit
+                    </a>
+                  )}
+                </td>
+                <td className={study.status.toLowerCase()}>{study.status}</td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
+      </div>
+    </div>
+  );
 };
 
 export default Dashboard;
