+++ frontend/src/pages/DummyStudies.jsx
