+++ frontend/src/pages/styles/ResultsPage.css
